+++ Services/Stores/Stores.Application/Services/CategoryService.cs
@@ -132,6 +132,14 @@ public class CategoryService : ICategoryService
         try
         {
             var category = await _repository.GetAsync(s => s.Id == categoryDto.Id, tracked: false);
+
+            if (category is null)
+            {
+                response.IsSuccessful = false;
+                response.Message = "Category not found!";
+                
+                return response;
+            }
             
             if (category.ConcurrencyStamp != categoryDto.ConcurrencyStamp)
             {
@@ -173,7 +181,52 @@ public class CategoryService : ICategoryService
         {
             var category = await _repository.GetAsync(s => s.Id == cateId, tracked: false);
             
+            if (category is null)
+            {
+                response.IsSuccessful = false;
+                response.Message = "Category not found!";
+                
+                return response;
+            }
+            
             await _repository.RemoveAsync(category);
+            
+            response.Body = _mapper.Map<CategoryDto>(category);
+            
+            return response;
+        }
+        catch (Exception ex)
+        {
+            response.IsSuccessful = false;
+            response.Message = ex.Message;
+            
+            return response;
+        }
+    }
+    
+    /// <summary>
+    /// Deleted (change state) the category
+    /// </summary>
+    /// <param name="cateId">The category's ID to delete</param>
+    /// <returns>The deleted category</returns>
+    public async Task<Response> DeleteAsync(Guid cateId)
+    {
+        var response = new Response();
+
+        try
+        {
+            var category = await _repository.GetAsync(s => s.Id == cateId, tracked: false);
+            
+            if (category is null)
+            {
+                response.IsSuccessful = false;
+                response.Message = "Category not found!";
+                
+                return response;
+            }
+
+            category.State = CategoryState.Deleted;
+            await _repository.UpdateAsync(category);
             
             response.Body = _mapper.Map<CategoryDto>(category);
