+++ Services/Stores/Stores.Application/Services/SubCategoryService.cs
@@ -19,16 +19,18 @@ public class SubCategoryService : ISubCategoryService
     /// <summary>
     /// Get list of subCategories
     /// </summary>
-    /// <param name="cateId">The main category that the sub-categories belong to</param>
-    /// <param name="pageSize">Pages number to get subCategories</param>
-    /// <param name="pageNumber">Page number to start with</param>
+    /// <param name="request">The main category that the sub-categories belong to</param>
     /// <returns>The subCategories list</returns>
-    public async Task<Response> GetAllAsync(Guid cateId, int pageSize = 0, int pageNumber = 1)
+    public async Task<Response> GetAllByCategoryIdAsync(GetSubCategoriesRequest request)
     {
         var response = new Response();
 
         try
         {
+            var cateId = request.CategoryId;
+            var pageSize = request.PageSize;
+            var pageNumber = request.PageNumber;
+            
             var subCategories = await _subCategoryRepository.GetAllAsync(
                 filter: s => s.CategoryId == cateId, 
                 tracked: false,
@@ -56,16 +58,18 @@ public class SubCategoryService : ISubCategoryService
     /// <summary>
     /// Get list of subCategories
     /// </summary>
-    /// <param name="cateName">The main category that the sub-categories belong to</param>
-    /// <param name="pageSize">Pages number to get subCategories</param>
-    /// <param name="pageNumber">Page number to start with</param>
+    /// <param name="request">The main category that the sub-categories belong to</param>
     /// <returns>The subCategories list</returns>
-    public async Task<Response> GetAllByCodeNameAsync(string cateName, int pageSize = 0, int pageNumber = 1)
+    public async Task<Response> GetAllByCategoryCodeNameAsync(GetSubCategoriesRequest request)
     {
         var response = new Response();
 
         try
         {
+            var cateName = request.CategoryName;
+            var pageSize = request.PageSize;
+            var pageNumber = request.PageNumber;
+            
             var subCategories = await _subCategoryRepository.GetAllAsync(
                 filter: s => s.Category!.CodeName == cateName, 
                 tracked: false,
@@ -102,6 +106,14 @@ public class SubCategoryService : ISubCategoryService
         try
         {
             var subCategory = await _subCategoryRepository.GetAsync(s => s.Id == subCateId, tracked: false);
+            
+            if (subCategory is null)
+            {
+                response.IsSuccessful = false;
+                response.Message = "SubCategory not found!";
+                
+                return response;
+            }
 
             response.Body = _mapper.Map<CategoryDto>(subCategory);
 
@@ -127,7 +139,13 @@ public class SubCategoryService : ISubCategoryService
 
         try
         {
-            var subCategory = _mapper.Map<SubCategory>(request.SubCategoryDto);
+
+            var subCategory = new SubCategory
+            {
+                Id = Guid.NewGuid(),
+                Name = request.Name,
+                CodeName = request.CodeName
+            };
 
             subCategory.CategoryId = request.CategoryId;
             
@@ -151,25 +169,39 @@ public class SubCategoryService : ISubCategoryService
     /// <summary>
     /// Update the subCategory's metadata
     /// </summary>
-    /// <param name="subCategoryDto">The subCategory to update</param>
+    /// <param name="request">The subCategory to update</param>
     /// <returns>The updated subCategory</returns>
-    public async Task<Response> UpdateAsync(SubCategoryDto subCategoryDto)
+    public async Task<Response> UpdateAsync(UpdateSubCategoryRequest request)
     {
         var response = new Response();
 
         try
         {
-            var subCategory = await _subCategoryRepository.GetAsync(s => s.Id == subCategoryDto.Id, tracked: false);
+            var subCategory = 
+                await _subCategoryRepository.GetAsync(s => s.Id == request.Id, tracked: false);
 
-            if (subCategory.ConcurrencyStamp != subCategoryDto.ConcurrencyStamp)
+            if (subCategory is null)
+            {
+                response.IsSuccessful = false;
+                response.Message = "SubCategory not found!";
+                
+                return response;
+            }
+
+            if (subCategory.ConcurrencyStamp != request.ConcurrencyStamp)
             {
                 response.IsSuccessful = false;
                 response.Message = "Concurrency conflict! Data was modified by another user!";
 
                 return response;
             }
-
-            var subCategoryToUpdate = _mapper.Map<SubCategory>(subCategoryDto);
+            
+            var subCategoryToUpdate = new SubCategory
+            {
+                CategoryId = request.CategoryId,
+                Name = request.Name,
+                CodeName = request.CodeName
+            };
             
             subCategoryToUpdate.ConcurrencyStamp = Guid.NewGuid();
 
@@ -201,7 +233,52 @@ public class SubCategoryService : ISubCategoryService
         {
             var subCategory = await _subCategoryRepository.GetAsync(s => s.Id == subCateId, tracked: false);
             
+            if (subCategory is null)
+            {
+                response.IsSuccessful = false;
+                response.Message = "SubCategory not found!";
+                
+                return response;
+            }
+            
             await _subCategoryRepository.RemoveAsync(subCategory);
+
+            response.Body = _mapper.Map<SubCategoryDto>(subCategory);
+
+            return response;
+        }
+        catch (Exception ex)
+        {
+            response.IsSuccessful = false;
+            response.Message = ex.ToString();
+
+            return response;
+        }
+    }
+    
+    /// <summary>
+    /// Deleted (change state) the subCategory
+    /// </summary>
+    /// <param name="subCateId">The subCategory's ID to delete</param>
+    /// <returns>The deleted subCategory</returns>
+    public async Task<Response> DeleteAsync(Guid subCateId)
+    {
+        var response = new Response();
+
+        try
+        {
+            var subCategory = await _subCategoryRepository.GetAsync(s => s.Id == subCateId, tracked: false);
+            
+            if (subCategory is null)
+            {
+                response.IsSuccessful = false;
+                response.Message = "SubCategory not found!";
+                
+                return response;
+            }
+
+            subCategory.State = CategoryState.Deleted;
+            await _subCategoryRepository.UpdateAsync(subCategory);
 
             response.Body = _mapper.Map<SubCategoryDto>(subCategory);
